#890: moves ImageSimilarityScorer to iped-engine module
